+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## CHANGELOG
 
+## 10/08/2022
+
+- [feature] Viết hàm đọc file json tại `fileUtils.go`
+- [feature] Bổ sung struct config file
+
 ## 09/08/2022
 
 - [feature] Bổ sung file config